refactor: rename `types/oauth.ts` to `types/oauth-external.ts`
